Fix dive computer copying
When copying a dive, we need to copy all the dive computers, but the first
one is special since it is embedded inside the "struct dive".

We had special case code for that case, but it only copied the samples and
events.

Use the proper "copy_dc()" for the primary dive computer copying too,
which gets all the details right.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>